+++ src/main/java/com/example/demo/exception/IExceptionHandlerBusinessException.java
@@ -1,0 +1,11 @@
+package com.example.demo.exception;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Component;
+
+@Component
+public interface IExceptionHandlerBusinessException {
+
+    ResponseEntity<?> handlerBusinessException(BusinessException e);
+
+}
